Moving checkProduct from fetchers to actions because env shouldn't be available to client component. Alternatively use server could be added to fetchers/product.ts
